requirements.txt
acres
aiofiles
aiohappyeyeballs
aiohttp
aiolimiter
aiosignal
annotated-types
anyio
asttokens
attrs
beautifulsoup4
biopython
boto3
botocore
bs4
cachetools
certifi
charset-normalizer
ci-info
click
colorama
configobj
configparser
dataclasses-json
db-dtypes
decorator
etelemetry
executing
faiss-cpu
filelock
frozenlist
google-api-core
google-auth
google-cloud
google-cloud-bigquery
google-cloud-core
google-cloud-secret-manager
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpc-google-iam-v1
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
httpx-sse
idna
iniconfig
ipython
ipython_pygments_lexers
isodate
itsdangerous
jedi
jmespath
joblib
jsonpatch
jsonpointer
langchain
langchain-aws
langchain-community
langchain-core
langchain-text-splitters
langchain-voyageai
langsmith
looseversion
lxml
marshmallow
matplotlib-inline
multidict
mypy-extensions
networkx
nibabel
nipype
numpy
orjson
packaging
pandas
parso
pathlib
pluggy
prompt_toolkit
propcache
proto-plus
protobuf
prov
pure_eval
puremagic
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
pydot
Pygments
PyMuPDF
pyparsing
PyPDF2
pytest
python-dateutil
python-dotenv
pytils
pytz
pyxnat
PyYAML
RapidFuzz
rdflib
requests
requests-toolbelt
rsa
s3transfer
scikit-learn
scipy
simplejson
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
tenacity
thefuzz
tools
torch
traitlets
traits
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
voyageai
wcwidth
yarl
zstandard
